refactor: rename new button
