+++ src/frontend/src/lib/components/core/Menu.svelte
@@ -9,7 +9,7 @@
 	import IconGitHub from '$lib/components/icons/IconGitHub.svelte';
 	import IconSettings from '$lib/components/icons/IconSettings.svelte';
 	import IconUser from '$lib/components/icons/IconUser.svelte';
-	import ButtonAsIcon from '$lib/components/ui/ButtonAsIcon.svelte';
+	import ButtonIcon from '$lib/components/ui/ButtonIcon.svelte';
 	import ButtonMenu from '$lib/components/ui/ButtonMenu.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import Hr from '$lib/components/ui/Hr.svelte';
@@ -37,14 +37,14 @@
 	$: walletOptions = !settingsRoute;
 </script>
 
-<ButtonAsIcon
+<ButtonIcon
 	bind:button
 	on:click={() => (visible = true)}
 	ariaLabel={$i18n.navigation.alt.menu}
 	testId={NAVIGATION_MENU_BUTTON}
 >
 	<IconUser />
-</ButtonAsIcon>
+</ButtonIcon>
 
 <Popover bind:visible anchor={button} direction="rtl">
 	<div class="flex flex-col gap-4" data-tid={NAVIGATION_MENU}>
