+++ src/frontend/src/lib/components/ui/ButtonIcon.svelte
